fix: fully fixed invalid field descriptors
This commit had fixed the invalid "Lunk;" (maybe represents UNKNOWN in short) field descriptors produced by old version of this program.

How it works?
We just read the full yarn mapping, and copy the descriptors from it to the final mapping.
But we had to deal with some classes that aren't fully deobfuscated, luckily we did it.